# docs/swagger.yaml
basePath: /
definitions:
...
  resource.InputPhoto:
    properties:
      caption:
        example: This is my greate team
        type: string
      photo_url:
        example: https://static.wikia.nocookie.net/naruto/images/5/50/Team_Kakashi.png/revision/latest?cb=20161219035928
        type: string
      title:
        example: Selfie with team
        type: string
    required:
    - photo_url
    - title
    type: object
...
info:
  contact:
    email: kadekbintanga@gmail.com
    name: API Support
  description: This is a service for final project
  license:
    name: Apache 2.0
  termsOfService: http://swagger.io/terms/
  title: Final Project by Bintang
  version: "1.0"
paths:
  /api/v1/allphoto:
    get:
      description: Get all photo  without any input
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all  photo handler
      tags:
      - Photo
...
  /api/v1/photo:
    get:
      description: Get photo with bearer token
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user photo
      tags:
      - Photo
    post:
      consumes:
      - application/json
      description: Create a photo with the input payload
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.InputPhoto'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a photo
      tags:
      - Photo
  /api/v1/photo/{photo_id}:
    delete:
      consumes:
      - application/json
      description: delete a photo with the token
      parameters:
      - description: Photo ID
        in: path
        name: photo_id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a photo
      tags:
      - Photo
    put:
      consumes:
      - application/json
      description: Update a photo with the input payload
      parameters:
      - description: Photo ID
        in: path
        name: photo_id
        required: true
        type: string
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.InputPhoto'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a photo
      tags:
      - Photo
...
swagger: "2.0"